requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.0
defusedxml==0.7.1
diffusers==0.32.2
distro==1.9.0
dj-database-url==0.5.0
Django==5.1.4
django-allauth==0.50.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
filelock==3.17.0
fsspec==2025.2.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
joblib==1.4.2
keybert==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
zipp==3.21.0
